refactor: ui caps
